+++ CHANGELOG.rst
@@ -1,7 +1,13 @@
 Changelog
 =========
 
-Release 1.0.1.0 (Development)
+Release 1.0.2.0 (Development)
 ---------------------------
+
+* Added control for mask overlap in `mnxc2`. `mnxc2` now passes the same tests as reference implementation.
+* Added an analog of scikit-image's `register_translation`, `masked_register_translation`. It will eventually replace `diff_register`.
+
+Release 1.0.1.0
+---------------
 
 * Added support for Gatan Digital Micrograph image formats DM3 and DM4